Update CONTRIBUTING.md
It's a really tiny issue, but "hitting" and "hacking" are words with slight violent connotations. It's not something that causes a huge issue, just something that may cause unconscious bias to a reader. Sounds dumb but if language that has that kind of connotation can be avoided I think it should be. Plus you had a typo in the word "without"